Fixed bug 219702.
The issue was that the the state in PostBase was set to AuthPending even if
first error recovery needed to be done. The pointer to the AuthBase class was
being initialized only when the authentication check was started. If the
plugin was asked to stop after error recovery request was made, but before the
authentication was started, then we used to assert because of null AuthBase
pointer.

This is a very rare crash, but needs to be fixed. To fix, a new enum was added
to indicate that the PostBase class is in error fixing state. The rest of the
changes are done to ensure that states are changed correctly.