When the file to be shared is deleted, it's entries from tracker get removed. Now, when the upload gets scheduled, the plugin tries to create a WebUpload::Entry class from the given xml file. The init of WebUpload::Media instance corresponding to the deleted file fails. We do not give an error there, and try to continue with the init'ing of any other files to be shared (which is 0 in this case). Now, when we try to do the actual upload, we see there is no WebUpload::Media which has not been sent, and set state as upload done.
Changes done:
(1) Tracker reading is done *AFTER* rest of the information is read from the
XML file
(2) Getting mime type, and original file path from tracker has been made
optional. So, if the file is deleted before it's upload is scheduled, then
not getting these does not indicate an error.
(3) Tracker mime type is used only if xml did not contain mime information
(4) With the above changes, a WebUpload::Media init will fail only if either
the XML file with the upload information, or tracker database is corrupted. In
either case, we should not continue with init of the WebUpload::Entry.
(5) Interaction with transfer-ui changed to handle the scenario where original
file link is not present.