Added handling of the signal emitted saying "about to unmount drive" just before device enters mass-storage mode.
Current limitation: If some other application blocks the device from entering
the mass storage mode, the current transfer enters into a blocked state. The
only ways out are to either somehow get the device into mass storage and then
exit it, or to cancel that transfer. This is an acceptable limitation, since,
sometime in the future, any application that blocks device entering
mass-storage mode will be killed, so that this scenario should not arise then.